feat(187) :: Generatation을 tech_course가 아닌 Student에 기록, Generatation 타입이 Long에서 Integer로 변경

+++ src/main/java/org/plteco/ploytechcourse/application/document/service/DocumentServiceApplicationImpl.java
@@ -89,7 +89,7 @@ public class DocumentServiceApplicationImpl implements DocumentServiceApplicatio
                 .map(HashTag::getName)
                 .toList();
 
-        Long generation = documentService.getUserGeneration(document).orElse(null);
+        Integer generation = documentService.getUserGeneration(document).orElse(null);
 
         String generationStr = generation == null ? "멘토" : String.valueOf(generation);
